==> notebooks/01_data_exploration.ipynb <==
import pandas as pd

df = pd.read_csv('../amazon_sales.csv')

# Load and view the data
print(df.head())
print(df.info())
print(df.describe())

# List of important categorical columns
categorical_cols = ['product_category','is_best_seller','is_sponsored','has_coupon','buy_box_availability']

# Explore each categorical column
for col in categorical_cols:
    print(f"Column: {col}")
    print(f"Number of unique values: {df[col].nunique()}")
    print("Top 10 most frequent values:")
    print(df[col].value_counts().head(10))
    print("Missing values:", df[col].isnull().sum())
    print("-"*50)

import matplotlib.pyplot as plt
import seaborn as sns

numeric_cols = ['product_rating','total_reviews','purchased_last_month','discounted_price','original_price','discount_percentage']

for col in numeric_cols:
    print(f"Summary statistics for {col}:")
    print(df[col].describe())
    print("Missing values:", df[col].isnull().sum())
    
    plt.figure(figsize=(10,4))
    sns.boxplot(x=df[col])
    plt.title(f"Boxplot of {col}")
    plt.show()
    
    plt.figure(figsize=(10,4))
    sns.histplot(df[col], bins=50, kde=True)
    plt.title(f"Histogram of {col}")
    plt.show()
